--- LittleLemonApp/Utility.js
@@ -1,8 +1,0 @@
-function transStateMap(arr){
-    return {
-        value: arr[0],
-        setValue: arr[1]
-    }
-}
-
-
